feat: Remove DocOps Writers from CODEOWNERS
This commit removes the writing half of the DocOps team from CODEOWNERS.
It is primarily engineers that iterate on the theme: this change ensures
that writers are not engaged with any in-flight work prematurely.

User testing will be explicitly requested when new elements or
functionality have reached an appropriate state of development.