Coveo: Fix vertical overflow race condition
Noticed that on hard page refreshes, and slower connections
that I would still see the extra vertical overflow caused by coveo.
I assumed it was a race condition, and wrapping the removal of
`atomic-relevance-inspector` in a timeout, would generally resolve it.
I didn't want a timeout to be the "fix" here, so I added the Mutation
Observer (https://developer.mozilla.org/en-US/docs/Web/API/MutationObserver)
which appears to be the safest approach.
There's an overall 5 second timeout still, in case search never renders
for whatever reason.